Add missing 'DROP INDEX' statements for upgrade files
When dropping a key in Oracle, one needs to specifically
request for the index to also be dropped.  Failing to do
that fails any tentative to recreate that primary key
with new fields.  This fixes that for the entitlements
tables in the previous commit adding entitlement id.

Change-Id: I365e0c7134097325e52178bb0e9c4259b4f0cb44
Addresses-Issue: Jira/AQUILON-5210
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Gaetan Menat <Gaetan.Menat@morganstanley.com>