aqdb: fix SQL statements for the operating_system DB change
The change was introduced in fcf720a84a443639a0a6245c28a07934321451ba
and contains .sql and .psql files talking about the "os" table
instead of the "operating_system" table.  This is fixed, as well
as the fact that Oracle complains if you repeat the "NOT NULL"
statement on something already "NOT NULL".

Change-Id: I719d7e3beff87fc82328ba956f6a409dd01c381d
Addresses-Issue: Jira/ECSETI-41
Reviewed-by: Gaetan Menat <Gaetan.Menat@morganstanley.com>
Tested-by: Aquilon Template Build testing and verification