Prevent use of wrong DNS domains for buildings
This adds a facility that allows internal components to check if a
given DNS domain is not already assigned to any other location
(including the building class) as its default DNS domain.
It also adds a new option '--force_dns_domain' to 'aq add_host', and
changes the default behaviour to disallow the use of FQDNs (specified
either with --hostname, or --prefix and --dns_domain) in DNS domains
already used as default DNS domains by buildings other than the one in
which the host being added is located.  The user can manually override
the new behaviour using the aforementioned new option.
New tests are included, and some old tests modified so that they take
into account the new behaviour.

Change-Id: I40f8a50d98959f165d8f44e4ea1b1ebe225ba5f7
Addresses-Issue: Jira/AQUILON-2479
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>