Add test to ensure duplicate list entries rejected
This adds a new regression test to verify that the broker reports an
error to the user when a file provided for the "aq reconfigure --list"
command contains duplicate entries.  The error message must include the
the following: "Provided list contains duplicate entry: %s", where %s is
the first of the detected duplicates.

Change-Id: I18c8adc1cccb6680dfd66bd4784103f3684bc44b
Addresses-Issue: Jira/AQUILON-616
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>