Make MockHub usable just after test_start
This ensures that the only dependency for tests that use MockHub is
test_start (i.e. it is now possible to use create isolated functional /
system tests independent of any (other than the non-optional test_start)
other tests, and the order in which the test harness executes them).
Apart from simplifying isolation and independence, this brings us one
step closer to the automatic parallelisation of tests.

Change-Id: I7621de75f7953066b6196383e56a6554a244fc4a
Addresses-Issue: Jira/AQUILON-6296
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>