Prevent deletion of countries containing cities
This ensures that a country that still contains at least one city cannot
be deleted using 'aq del_country' irrespective of whether the city is a
direct child, or an indirect descendant contained in a campus that
belongs to the country.
This also adds a new option '--force_non_empty' which allows the user to
delete countries that still contain one or more cities.
This also adds a new useful method to the Location class, which allows
us to perform such containment tests for objects of type Location, or
any of its subtypes.
Last but not least, method MockHub.delete_countries has been modify to
unconditionally delete all countries.

Change-Id: I72ae4b5141d1994ce37c94f7d0649d412b645534
Addresses-Issue: Jira/AQUILON-6192
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>