New option to report objects in-scope for CM
This adds a new option --cm_check to all commands that potentially
require change management.  When present, the new option causes such
commands to stop at ChangeManagement.validate(), and raise an exception
to return the list of all in-scope objects (that would be affected by
the command if it completed) to the user.  All commands that invoke the
consider() and validate() methods of the ChangeManagement class (usually
in their render() methods) are affected.
This also adds information about the --resume option to the message that
prompts the user to restart tests after failure if runtests.py was
executed with the --failfast option.

Change-Id: Ibae4bc02a97d41cca17ca5cfe14f1526b5f29b38
Addresses-Issue: Jira/AQUILON-4508
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>