Change how del_host and buildstatus interact
This ensures that, by default, del_host does not delete hosts which
are in states other than blind, build, decommissioned, failed or
install.
This also adds a new configuration option which allows archetypes to
be configured to override the new default behaviour described above.
File aqd.conf.defaults is modified to ensure that the new default
applies to archetype 'aquilon', and that del_host can still be used as
before (i.e. to delete hosts irrespective of their build state) for the
following archetypes: aurora, esx_cluster, gridcluster, hacluster,
metacluster, netinfra, storagecluster, vmhost, and windows.
This also includes a couple of small improvements to MockHub (used in
the new tests for del_host).

Change-Id: I075573f3d9730a9bdad342e510b4e14ea56116d0
Addresses-Issue: Jira/AQUILON-6292
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>