Include ownership EON IDs in change-management log
This ensures that ownership EON IDs are included in the
change-management log.
This also modifies class MockHub so that preexisting GRNs, their
associated EON IDs, and their disabled status are preserved after an
instance of MockHub is deleted.  All GRNs created between the creation
of a MockHub instance, and its deletion, are also deleted.

Change-Id: I171fd14239ecb6d09f4b1853a60729af5e44f5aa
Addresses-Issue: Jira/AQUILON-6253
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>