Allow service address PTR to point at shared name
This adds an option --map_to_shared_name option for service addition and
update that causes the reverse PTR DNS record to point back to the FQDN
of the shared-name (within a resource-group).  If the 'sa_aliases' flag
is set for the shared-name, an address alias from the shared-name to the
service address is created (and removed on service address deletion).

Change-Id: Icdf078985321a5d035f3d5ac41a45b3f1e4d2393
Addresses-Issue: Jira/AQUILON-6374
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Tomasz Kotarba <Tomasz.Kotarba@morganstanley.com>