New option --default_dns_domain_from
This modifies the location wrapper used to compute a default DNS domain
for a given location so that it now allows for all objects used to
compute such a domain to be filtered out unless they belong to a given
class (e.g. Hub).
This also adds new tests, a new option --default_dns_domain_from to
command add_service_address, and adds a missing requirement for
option --dns_domain which can now also be used together with
--shortname (instead of just --prefix).

Change-Id: I35656375f5d06054b7f9af3a607081aebee24700
Addresses-Issue: Jira/AQUILON-6621
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>